Try to load lz4-java from java.library.path, then fallback to bundled
Use system-provided lz4-java, if it's available on the java.library.path
(the default location where Java looks for native libraries), and fallback
to bundled binary if it's not. This allows user to use lz4 java even
on systems that are not directly supported, like Linux with musl libc,
uClibc etc.