Allow obtaining and restoring the streaming hashes' internal state.
This enables a "very large stream" hashing use case: Instead of trying
to read a large object in a single pass, one can fetch chunks of it at
a time and save the state after each chunk as a checkpoint. In the case
of a failure, the saved checkpoint can be used to retry from that point
of the stream onwards.