support bytebuffer updates on streaming hashes
In some cases ByteBuffers do not provide access to an underlying
byte array without incurring copies, this is notably the case
for off-heap direct bytebuffers.

This patch provides a way to call update on streaming XXHASH instances
against a ByteBuffer, which avoids the extra copy needed into a byte
array.